pyproject.toml
[project]
name = "finllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "beautifulsoup4>=4.14.3",
    "datasets>=4.7.0",
    "evaluate>=0.4.6",
    "html5lib>=1.1",
    "huggingface>=0.0.1",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "peft>=0.18.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sec-edgar-downloader>=5.1.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "google-genai>=1.0.0",
    "sentence-transformers>=3.0.0",
    "requests>=2.31.0",
    "pymupdf>=1.24.0",
    "ollama>=0.4.0",
    "chromadb>=0.5.0",
]
